+++ src/marine_qc/helpers/spherical_geometry.py
@@ -60,8 +60,8 @@ def _geod_inv(
     return fwd_az, back_az, dist
 
 
-@post_format_return_type(["lat1", "lon1", "lat2", "lon2"], dtype=float)
-@inspect_arrays(["lat1", "lon1", "lat2", "lon2"])
+@post_format_return_type("lat1", "lon1", "lat2", "lon2", dtype=float)
+@inspect_arrays("lat1", "lon1", "lat2", "lon2")
 def angular_distance(
     lat1: SequenceNumberType,
     lon1: SequenceNumberType,
@@ -106,8 +106,8 @@ def angular_distance(
     return result
 
 
-@post_format_return_type(["lat1", "lon1", "lat2", "lon2"], dtype=float)
-@inspect_arrays(["lat1", "lon1", "lat2", "lon2"])
+@post_format_return_type("lat1", "lon1", "lat2", "lon2", dtype=float)
+@inspect_arrays("lat1", "lon1", "lat2", "lon2")
 def sphere_distance(
     lat1: SequenceNumberType,
     lon1: SequenceNumberType,
@@ -156,8 +156,8 @@ def sphere_distance(
     return result
 
 
-@post_format_return_type(["lat1", "lon1", "lat2", "lon2", "f"], dtype=float, multiple=True)
-@inspect_arrays(["lat1", "lon1", "lat2", "lon2", "f"])
+@post_format_return_type("lat1", "lon1", "lat2", "lon2", "f", dtype=float, multiple=True)
+@inspect_arrays("lat1", "lon1", "lat2", "lon2", "f")
 def intermediate_point(
     lat1: SequenceNumberType,
     lon1: SequenceNumberType,
@@ -214,8 +214,8 @@ def intermediate_point(
     return lat_f, lon_f
 
 
-@post_format_return_type(["lat1", "lon1", "lat2", "lon2"], dtype=float)
-@inspect_arrays(["lat1", "lon1", "lat2", "lon2"])
+@post_format_return_type("lat1", "lon1", "lat2", "lon2", dtype=float)
+@inspect_arrays("lat1", "lon1", "lat2", "lon2")
 def course_between_points(
     lat1: SequenceNumberType,
     lon1: SequenceNumberType,
@@ -253,8 +253,8 @@ def course_between_points(
     return fwd_azimuth
 
 
-@post_format_return_type(["lat1", "lon1"], dtype=float, multiple=True)
-@inspect_arrays(["lat1", "lon1"])
+@post_format_return_type("lat1", "lon1", dtype=float, multiple=True)
+@inspect_arrays("lat1", "lon1")
 def lat_lon_from_course_and_distance(
     lat1: SequenceNumberType,
     lon1: SequenceNumberType,